requirements.txt
amqp==5.3.1
annotated-types==0.7.0
APIFlask==3.0.2
apispec==6.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
billiard==4.2.4
blinker==1.9.0
celery==5.6.2
cffi==2.0.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.13.3
Deprecated==1.3.1
dnspython==2.8.0
email-validator==2.3.0
fakeredis==2.33.0
fakeredis-fix==0.4.1
Flask==3.1.2
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==4.1.1
flask-marshmallow==1.3.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.6.2
limits==5.8.0
MarkupSafe==3.0.3
marshmallow==4.2.1
ordered-set==4.1.0
packaging==26.0
pluggy==1.6.0
prompt_toolkit==3.0.52
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-cov==7.0.0
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis>=4.6,<5
six==1.17.0
sortedcontainers==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
webargs==8.7.1
Werkzeug==3.1.5
wrapt==2.1.1
beautifulsoup4
requests